quantum_secured_links/requirements.txt
numpy
pycryptodomex
